+++ tests/StoreTransportMessageIdTest.php
@@ -4,6 +4,7 @@ namespace Spatie\MailcoachSesFeedback\Tests;
 
 use Illuminate\Mail\Events\MessageSent;
 use Spatie\Mailcoach\Models\Send;
+use Spatie\MailcoachSesFeedback\Tests\factories\SendFactory;
 use Swift_Message;
 
 class StoreTransportMessageIdTest extends TestCase
@@ -11,7 +12,7 @@ class StoreTransportMessageIdTest extends TestCase
     /** @test **/
     public function it_stores_the_message_id_from_the_transport()
     {
-        $pendingSend = factory(Send::class)->create();
+        $pendingSend = (new SendFactory())->create();
         $message = new Swift_Message('Test', 'body');
         $message->getHeaders()->addTextHeader('X-Ses-Message-ID', '1234');